resonanceFit: move plotting functions to separate source file
In order to decrease the line count in 'resonanceFit.cc' move all
functions concerning the plotting to a separate source file.
